t300 Naming improvements
